[add] Copy methods of Lark documents [add] independent Drive & Document Block modules [optimize] update Upstream packages
