+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "mobx-lark",
-    "version": "2.5.0",
+    "version": "2.6.0-rc.0",
     "license": "LGPL-3.0",
     "author": "shiy2008@gmail.com",
     "description": "Unofficial TypeScript SDK for FeiShu/Lark API, which is based on MobX-RESTful.",
@@ -25,37 +25,37 @@
         "url": "https://github.com/idea2app/MobX-Lark/issues"
     },
     "dependencies": {
-        "@swc/helpers": "^0.5.17",
-        "@types/react": "^19.2.2",
+        "@swc/helpers": "^0.5.18",
+        "@types/react": "^19.2.7",
         "koajax": "^3.1.2",
         "lodash.memoize": "^4.1.2",
         "mobx": "^6.15.0",
         "mobx-restful": "^2.1.4",
         "regenerator-runtime": "^0.14.1",
-        "web-utility": "^4.6.2"
+        "web-utility": "^4.6.4"
     },
     "peerDependencies": {
         "react": ">=16"
     },
     "devDependencies": {
-        "@parcel/config-default": "~2.16.0",
-        "@parcel/transformer-typescript-tsc": "~2.16.0",
+        "@parcel/config-default": "~2.16.3",
+        "@parcel/transformer-typescript-tsc": "~2.16.3",
         "@types/fs-extra": "^11.0.4",
         "@types/lodash.memoize": "^4.1.9",
-        "@types/node": "^22.18.12",
-        "@types/react-dom": "^19.2.2",
-        "@types/turndown": "^5.0.5",
+        "@types/node": "^22.19.3",
+        "@types/react-dom": "^19.2.3",
+        "@types/turndown": "^5.0.6",
         "dotenv": "^17.2.3",
-        "fs-extra": "^11.3.2",
+        "fs-extra": "^11.3.3",
         "husky": "^9.1.7",
-        "lint-staged": "^16.2.5",
-        "parcel": "~2.16.0",
-        "prettier": "^3.6.2",
-        "react": "^19.2.0",
-        "react-dom": "^19.2.0",
+        "lint-staged": "^16.2.7",
+        "parcel": "~2.16.3",
+        "prettier": "^3.7.4",
+        "react": "^19.2.3",
+        "react-dom": "^19.2.3",
         "ts-node": "^10.9.2",
-        "turndown": "^7.2.1",
-        "typedoc": "^0.28.14",
+        "turndown": "^7.2.2",
+        "typedoc": "^0.28.15",
         "typedoc-plugin-mdn-links": "^5.0.10",
         "typescript": "~5.8.3"
     },
